requirements.txt
beautifulsoup4==4.5.1
Django==1.10.2
django-appconf==1.0.2
django-channels==0.7.0
django-compressor==2.1
django-filebrowser==3.7.2
django-grappelli==2.8.2
django-model-utils==2.6
django-tinymce==2.4.0
Django-tinymce-filebrowser==0.1.1
django-user-sessions==1.3.1
gunicorn==19.6.0
leveldb==0.194
lmdb==0.92
oauthlib==2.0.0
Pillow==3.4.2
psycopg2==2.6.2
python-dateutil==2.5.3
python-ldap==2.4.27
rcssmin==1.0.6
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
typing==3.5.2.2
